+++ modules/hac/exec/src/sap/commerce/toolset/hac/exec/http/AuthContextCache.kt
@@ -23,8 +23,8 @@ import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import sap.commerce.toolset.exec.context.ReplicaContext
-import sap.commerce.toolset.exec.settings.state.ConnectionSettingsState
 import sap.commerce.toolset.hac.exec.settings.event.HacConnectionSettingsListener
+import sap.commerce.toolset.hac.exec.settings.state.AuthMode
 import sap.commerce.toolset.hac.exec.settings.state.HacConnectionSettingsState
 import java.util.concurrent.ConcurrentHashMap
 
@@ -36,20 +36,21 @@ class AuthContextCache(private val project: Project) : Disposable {
     init {
         project.messageBus.connect().subscribe(HacConnectionSettingsListener.TOPIC, object : HacConnectionSettingsListener {
             override fun onDelete(connection: HacConnectionSettingsState) = invalidateCookies(connection)
-            override fun onUpdate(settings: Collection<HacConnectionSettingsState>) = settings.forEach { invalidateCookies(it) }
+            override fun onUpdate(settings: Collection<HacConnectionSettingsState>) = settings
+                .filterNot { it.authMode == AuthMode.MANUAL }
+                .forEach { invalidateCookies(it) }
+
             override fun onSave(settings: Collection<HacConnectionSettingsState>) = settings.forEach { invalidateCookies(it) }
         })
     }
 
     override fun dispose() = authContexts.clear()
 
-    fun getKey(settings: ConnectionSettingsState, context: ReplicaContext? = null) = "${settings.uuid}_${context?.replicaId ?: "auto"}"
+    fun getKey(settings: HacConnectionSettingsState, context: ReplicaContext? = null) = "${settings.uuid}_${context?.replicaId ?: "auto"}"
 
-    private fun invalidateCookies(settings: ConnectionSettingsState) {
-        authContexts.keys
-            .filter { it.startsWith(settings.uuid) }
-            .forEach { authContexts.remove(it) }
-    }
+    private fun invalidateCookies(settings: HacConnectionSettingsState) = authContexts.keys
+        .filter { it.startsWith(settings.uuid) }
+        .forEach { authContexts.remove(it) }
 
     companion object {
         fun getInstance(project: Project): AuthContextCache = project.service()