+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [2025.2.4.6]
 
 <cite>Release contributors</cite>
-- 20 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.6+author%3Amlytvyn+is%3Apr)
+- 21 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.6+author%3Amlytvyn+is%3Apr)
 - 1 PR(s) by [Eugeni_Kalenchuk](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.6+author%3Aekalenchuk+is%3Apr)
 
 ### `CCv2` enhancements
@@ -37,7 +37,8 @@
 - Display website access errors in the Browser for `manual` authentication mode [#1641](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1641)
 - Added `Got it Tooltip` describing new authentication modes [#1651](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1651)
 - Adjusted UI of the dialog to fit the content better [#1655](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1655)
-- Added proxy authorization for hAC [#1656](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1656)
+- Added proxy authorization for `automatic` authentication mode [#1656](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1656)
+- Improved manual authentication mode [#1657](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1657)
 
 ## [2025.2.4.5]